+++ demo/cypress/index.ts
@@ -15,3 +15,4 @@ export { Wizard } from './Wizard';
 export { DropDown } from './DropDown';
 export { RadioGroup } from './RadioGroup';
 export { MultiSelect } from './MultiSelect';
+export { NumericTextBox } from './NumericTextBox';