+++ CHANGELOG.md
@@ -5,8 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [0.33.0] - 2020-10-26
-
-
+- NumericTextBox: Added dynamic validation by min/max 
 
 ## [0.32.0] - 2020-10-12